IConsole replaces Console.
