--- Console/CMakeLists.txt
@@ -1,42 +1,0 @@
-#   Copyright 2014 Wolfgang Thaller.
-#
-#   This file is part of Retro68.
-#
-#   Retro68 is free software: you can redistribute it and/or modify
-#   it under the terms of the GNU General Public License as published by
-#   the Free Software Foundation, either version 3 of the License, or
-#   (at your option) any later version.
-#
-#   Retro68 is distributed in the hope that it will be useful,
-#   but WITHOUT ANY WARRANTY; without even the implied warranty of
-#   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#   GNU General Public License for more details.
-#
-#   You should have received a copy of the GNU General Public License
-#   along with Retro68.  If not, see <http://www.gnu.org/licenses/>.
-
-add_library(RetroConsole
-    retro/Console.cc
-    retro/Console.h
-    retro/ConsoleWindow.cc
-    retro/ConsoleWindow.h
-    retro/MacUtils.h
-    retro/InitConsole.cc
-    )
-set_target_properties(retrocrt
-    PROPERTIES
-        COMPILE_OPTIONS -ffunction-sections)
-
-# different library name for Carbon
-# (Carbon shares the powerpc-apple-macos/ directory with Classic PPC)
-if(CMAKE_SYSTEM_NAME MATCHES RetroCarbon)
-    set_target_properties(RetroConsole PROPERTIES OUTPUT_NAME RetroConsoleCarbon)
-endif()
-target_include_directories(RetroConsole PUBLIC .)
-
-install(TARGETS RetroConsole  DESTINATION lib)
-
-add_application(ConsoleTest
-    ConsoleTest.cc
-    )
-target_link_libraries(ConsoleTest RetroConsole)
