+++ Console/ConsoleTest.cc
@@ -9,7 +9,7 @@ namespace retro
 int main()
 {
     retro::InitConsole();
-    std::string out = "Hello, world.\nEnter \"exit\" to quit.\n";
+    std::string out = "Hello, \033[1mexternal world of \033[0m\033[3mtrue beauty and \033[4mgreatness\033[0m.\nEnter \"exit\" to quit.\n";
     retro::Console::currentInstance->write(out.data(), out.size());
     
     std::string in;
