+++ Console/retro/Console.h
@@ -1,5 +1,5 @@
 /*
-    Copyright 2012 Wolfgang Thaller.
+    Copyright 2012-2020 Wolfgang Thaller, Davide Bucci
 
     This file is part of Retro68.
 
@@ -25,6 +25,39 @@
 
 namespace retro
 {
+    class Attributes
+    {
+    public:
+
+        bool isBold(void) const;
+        bool isUnderline(void) const;
+        bool isItalic(void) const;
+        
+        void setBold(const bool v);
+        void setUnderline(const bool v);
+        void setItalic(const bool v);
+        
+        Attributes(void);
+        void reset(void);
+
+    private:
+        
+        bool cBold;
+        bool cUnderline;
+        bool cItalic;
+    };
+    
+    class AttributedChar
+    {
+    public:
+        char c;
+        Attributes attrs;
+        AttributedChar(char cc, Attributes aa) {c=cc; attrs=aa;}
+    };
+
+//    inline bool operator==(const Attributes& lhs, const Attributes& rhs);
+//    inline bool operator!=(const Attributes& lhs, const Attributes& rhs);
+
     class Console
     {
     public:
@@ -52,8 +85,11 @@ namespace retro
     private:
         GrafPtr consolePort = nullptr;
         Rect bounds;
+        Attributes currentAttr;
 
-        std::vector<char> chars, onscreen;
+        std::vector<AttributedChar> chars, onscreen;
+        bool isProcessingEscSequence;
+        int sequenceStep;
 
         short cellSizeX;
         short cellSizeY;
@@ -72,10 +108,13 @@ namespace retro
         void PutCharNoUpdate(char c);
         void Update();
 
+        short CalcStartX(short x, short y);
         Rect CellRect(short x, short y);
         void DrawCell(short x, short y, bool erase = true);
         void DrawCells(short x1, short x2, short y, bool erase = true);
         void ScrollUp(short n = 1);
+        void ProcessEscSequence(char c);
+        void SetAttributes(Attributes aa);
         
         void InvalidateCursor();
