--- IConsole/retro/IConsoleWindow.cc
@@ -1,131 +1,0 @@
-/*
-    Copyright 2012 Wolfgang Thaller.
-
-    This file is part of Retro68.
-
-    Retro68 is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    Retro68 is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with Retro68.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#include "IConsoleWindow.h"
-#include "Events.h"
-#include <unordered_map>
-
-using namespace retro;
-
-namespace
-{
-    std::unordered_map<WindowPtr, IConsoleWindow*>    *windows = NULL;
-}
-
-IConsoleWindow::IConsoleWindow(Rect r, ConstStr255Param title)
-{
-    GrafPtr port;
-        //Retro68 Improved Console
-    win = NewWindow(NULL, &r, "\pThe Queen's Footsteps", true, 0, (WindowPtr)-1, true, 0);
-
-#if !TARGET_API_MAC_CARBON
-    port = win;
-    Rect portRect = port->portRect;
-#else
-    port = GetWindowPort(win);
-    Rect portRect;
-    GetPortBounds(port, &portRect);
-#endif
-
-    SetPort(port);
-    EraseRect(&portRect);
-
-    if(!windows)
-        windows = new std::unordered_map<WindowPtr, IConsoleWindow*>();
-    (*windows)[win] = this;
-
-    Init(port, portRect);
-}
-
-IConsoleWindow::~IConsoleWindow()
-{
-    windows->erase(win);
-    DisposeWindow(win);
-}
-
-char IConsoleWindow::WaitNextChar()
-{
-    EventRecord event;
-    WindowPtr eventWin;
-    IConsoleWindow *realConsole;
-#if TARGET_API_MAC_CARBON
-    Rect *boundsPtr = NULL;
-#else
-    Rect *boundsPtr = &qd.screenBits.bounds;
-#endif
-    
-    do
-    {
-        #if TARGET_API_MAC_CARBON
-            #define SystemTask()
-        #endif
-        SystemTask();
-        Idle();
-        while(!GetNextEvent(everyEvent, &event))
-        {
-            SystemTask();
-            Idle();
-        }
-            
-        switch(event.what)
-        {
-            case updateEvt:
-                eventWin = (WindowPtr)event.message;
-                realConsole = (*windows)[(WindowPtr)event.message];
-                if(realConsole)
-                {
-                    Rect updateRect;
-                    BeginUpdate(eventWin);
-#if TARGET_API_MAC_CARBON
-                    RgnHandle rgn = NewRgn();
-                    GetPortVisibleRegion(GetWindowPort(eventWin), rgn);
-                    GetRegionBounds(rgn, &updateRect);
-                    DisposeRgn(rgn);
-#else
-                    updateRect = (*qd.thePort->visRgn)->rgnBBox; // Life was simple back then.
-#endif
-                    realConsole->Draw(updateRect);
-                    EndUpdate(eventWin);
-                }
-                break;
-            case mouseDown:
-                switch(FindWindow(event.where, &eventWin))
-                {
-                    case inDrag:
-                        DragWindow(eventWin, event.where, boundsPtr);
-                        break;
-                    case inGrow:
-                        {
-                            long growResult = GrowWindow(eventWin, event.where, boundsPtr);
-                            SizeWindow(eventWin, growResult & 0xFFFF, growResult >> 16, false);
-                            Reshape(Rect {0, 0, (short) (growResult >> 16), (short) (growResult & 0xFFFF) });
-                        }
-                        break;
-                    case inGoAway:
-                        {
-                            if (TrackGoAway(eventWin,event.where))
-                                exit(0);
-                        }
-                }
-                break;
-        }
-    } while(event.what != keyDown && event.what != autoKey);
-    
-    return event.message & charCodeMask;
-}
